style: move runtime apply tests to their own module
1.5k lines of inline tests is wayy excessive.
